add cars (see desc.); logout(see desc.)
add cars for users now works but it submits values (like year) to the
database that the table doesn't like (year field is in date format).
Logout was nearly fixed, although it seems funky yet. Kyle's nav bar was
indeed backwards and now shows the right information according to
whether a user is logged in. loggedin was returning false because on the
actual loggedIn.php file, the require/include statement pointed to the
wrong filepath for the necessary functions.php